Extract interface for LucenePrimaryKeySegmentIndex
and update usages
